refactor: 객체와 기능 분리
